script/requirements.txt
# 服务器 python 版本 3.7
pyspark==3.2.4
bert4vec==0.3.0
pymilvus==2.1.3
esdk-obs-python==3.22.2